@@ -9,6 +9,6 @@
   "main": "build/index.js",
   "dependencies": {
     "jest-editor-support": "^21.2.0",
-    "typescript": "^2.2.1"
+    "typescript": "^2.5.3"
   }
 }
