@@ -10,8 +10,8 @@
 
 'use strict';
 
-const os = require('os');
-const docblock = require('../');
+import os from 'os';
+import * as docblock from '..';
 
 describe('docblock', () => {
   it('extracts valid docblock with line comment', () => {
@@ -211,5 +211,202 @@ describe('docblock', () => {
     expect(docblock.parse(code)).toEqual({
       providesModule: 'apple/banana',
     });
+  });
+
+  it('extracts comments from docblock', () => {
+    const code =
+      '/**' +
+      os.EOL +
+      ' * hello world' +
+      os.EOL +
+      ' * @flow yes' +
+      os.EOL +
+      ' */';
+    expect(docblock.parseWithComments(code)).toEqual({
+      comments: 'hello world',
+      pragmas: {flow: 'yes'},
+    });
+  });
+
+  it('extracts multiline comments from docblock', () => {
+    const code =
+      '/**' +
+      os.EOL +
+      ' * hello' +
+      os.EOL +
+      ' * world' +
+      os.EOL +
+      ' * @flow yes' +
+      os.EOL +
+      ' */';
+    expect(docblock.parseWithComments(code)).toEqual({
+      comments: 'hello' + os.EOL + 'world',
+      pragmas: {flow: 'yes'},
+    });
+  });
+
+  it('preserves leading whitespace in multiline comments from docblock', () => {
+    const code =
+      '/**' + os.EOL + ' *  hello' + os.EOL + ' *   world' + os.EOL + ' */';
+
+    expect(docblock.parseWithComments(code).comments).toEqual(
+      ' hello' + os.EOL + '  world',
+    );
+  });
+
+  it('extracts comments from beginning and end of docblock', () => {
+    const code =
+      '/**' +
+      os.EOL +
+      ' * hello' +
+      os.EOL +
+      ' * @flow yes' +
+      os.EOL +
+      ' * ' +
+      os.EOL +
+      ' * world' +
+      os.EOL +
+      ' */';
+    expect(docblock.parseWithComments(code)).toEqual({
+      comments: 'hello' + os.EOL + os.EOL + 'world',
+      pragmas: {flow: 'yes'},
+    });
+  });
+
+  it('extracts docblock comments as CRLF when docblock contains CRLF', () => {
+    const code = '/**\r\n * foo\r\n * bar\r\n*/';
+    expect(docblock.parseWithComments(code)).toEqual({
+      comments: 'foo\r\nbar',
+      pragmas: {},
+    });
+  });
+
+  it('extracts docblock comments as LF when docblock contains LF', () => {
+    const code = '/**\n * foo\n * bar\n*/';
+    expect(docblock.parseWithComments(code)).toEqual({
+      comments: 'foo\nbar',
+      pragmas: {},
+    });
+  });
+
+  it('strips the docblock out of a file that contains a top docblock', () => {
+    const code = '/**\n * foo\n * bar\n*/\nthe rest';
+    expect(docblock.strip(code)).toEqual('\nthe rest');
+  });
+
+  it('returns a file unchanged if there is no top docblock to strip', () => {
+    const code = 'someCodeAtTheTop();\n/** docblock */';
+    expect(docblock.strip(code)).toEqual(code);
+  });
+
+  it('prints docblocks with no pragmas as empty string', () => {
+    const pragmas = {};
+    expect(docblock.print({pragmas})).toEqual('');
+  });
+
+  it('prints docblocks with one pragma on one line', () => {
+    const pragmas = {flow: ''};
+    expect(docblock.print({pragmas})).toEqual('/** @flow */');
+  });
+
+  it('prints docblocks with multiple pragmas on multiple lines', () => {
+    const pragmas = {
+      flow: '',
+      format: '',
+    };
+    expect(docblock.print({pragmas})).toEqual(
+      '/**' + os.EOL + ' * @flow' + os.EOL + ' * @format' + os.EOL + ' */',
+    );
+  });
+
+  it('prints docblocks with pragmas', () => {
+    const pragmas = {
+      flow: 'foo',
+      providesModule: 'x/y/z',
+    };
+    expect(docblock.print({pragmas})).toEqual(
+      '/**' +
+        os.EOL +
+        ' * @flow foo' +
+        os.EOL +
+        ' * @providesModule x/y/z' +
+        os.EOL +
+        ' */',
+    );
+  });
+
+  it('prints docblocks with comments', () => {
+    const pragmas = {flow: 'foo'};
+    const comments = 'hello';
+    expect(docblock.print({comments, pragmas})).toEqual(
+      '/**' +
+        os.EOL +
+        ' * hello' +
+        os.EOL +
+        ' *' +
+        os.EOL +
+        ' * @flow foo' +
+        os.EOL +
+        ' */',
+    );
+  });
+
+  it('prints docblocks with comments and no keys', () => {
+    const pragmas = {};
+    const comments = 'Copyright 2004-present Facebook. All Rights Reserved.';
+    expect(docblock.print({comments, pragmas})).toEqual(
+      '/**' + os.EOL + ' * ' + comments + os.EOL + ' */',
+    );
+  });
+
+  it('prints docblocks with multiline comments', () => {
+    const pragmas = {};
+    const comments = 'hello' + os.EOL + 'world';
+    expect(docblock.print({comments, pragmas})).toEqual(
+      '/**' + os.EOL + ' * hello' + os.EOL + ' * world' + os.EOL + ' */',
+    );
+  });
+
+  it('prints docblocks that are parseable', () => {
+    const pragmas = {a: 'b', c: ''};
+    const comments = 'hello world!';
+    const formatted = docblock.print({comments, pragmas});
+    const parsed = docblock.parse(formatted);
+    expect(parsed).toEqual(pragmas);
+  });
+
+  it('can augment existing docblocks with comments', () => {
+    const before =
+      '/**' + os.EOL + ' * Legalese' + os.EOL + ' * @flow' + os.EOL + ' */';
+    const {comments, pragmas} = docblock.parseWithComments(before);
+    pragmas.format = '';
+    const after = docblock.print({comments, pragmas});
+    expect(after).toEqual(
+      '/**' +
+        os.EOL +
+        ' * Legalese' +
+        os.EOL +
+        ' *' +
+        os.EOL +
+        ' * @flow' +
+        os.EOL +
+        ' * @format' +
+        os.EOL +
+        ' */',
+    );
+  });
+
+  it('prints docblocks using CRLF if comments contains CRLF', () => {
+    const pragmas = {};
+    const comments = 'hello\r\nworld';
+    const formatted = docblock.print({comments, pragmas});
+    expect(formatted).toEqual('/**\r\n * hello\r\n * world\r\n */');
+  });
+
+  it('prints docblocks using LF if comments contains LF', () => {
+    const pragmas = {};
+    const comments = 'hello\nworld';
+    const formatted = docblock.print({comments, pragmas});
+    expect(formatted).toEqual('/**\n * hello\n * world\n */');
   });
 });
