@@ -5,11 +5,12 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {EventEmitter} from 'events';
+import { EventEmitter } from 'events';
 
 export class Runner extends EventEmitter {
   constructor(workspace: ProjectWorkspace);
-  start(): void;
+  watchMode: boolean;
+  start(watchMode?: boolean): void;
   closeProcess(): void;
   runJestWithUpdateForSnapshots(completion: any): void;
 }
@@ -28,7 +29,7 @@ export class ProjectWorkspace {
     rootPath: string,
     pathToJest: string,
     pathToConfig: string,
-    localJestMajorVersin: number
+    localJestMajorVersin: number,
   );
   pathToJest: string;
   rootPath: string;
@@ -61,15 +62,19 @@ export class Expect extends Node {}
 
 export class TestReconciler {
   stateForTestFile(file: string): TestReconcilationState;
-  stateForTestAssertion(file: string, name: string): TestFileAssertionStatus | null;
-  failedStatuses(): Array<TestFileAssertionStatus>;
-  updateFileWithJestStatus(data): void;
+  assertionsForTestFile(file: string): TestAssertionStatus[] | null;
+  stateForTestAssertion(
+    file: string,
+    name: string,
+  ): TestFileAssertionStatus | null;
+  updateFileWithJestStatus(data): TestFileAssertionStatus[];
 }
 
-export type TestReconcilationState = "Unknown" |
-  "KnownSuccess" |
-  "KnownFail" |
-  "KnownSkip";
+export type TestReconcilationState =
+  | 'Unknown'
+  | 'KnownSuccess'
+  | 'KnownFail'
+  | 'KnownSkip';
 
 export interface TestFileAssertionStatus {
   file: string;
@@ -91,7 +96,7 @@ export interface JestFileResults {
   name: string;
   summary: string;
   message: string;
-  status: "failed" | "passed";
+  status: 'failed' | 'passed';
   startTime: number;
   endTime: number;
   assertionResults: Array<JestAssertionResults>;
@@ -100,7 +105,7 @@ export interface JestFileResults {
 export interface JestAssertionResults {
   name: string;
   title: string;
-  status: "failed" | "passed";
+  status: 'failed' | 'passed';
   failureMessages: string[];
 }
 
