@@ -11,20 +11,24 @@ import {readFileSync} from 'fs';
 
 import {parse as babylonParse} from 'babylon';
 import {Expect, ItBlock} from './parser_nodes';
+import type {File as BabylonFile} from 'babylon';
 
 export type BabylonParserResult = {
   expects: Array<Expect>,
   itBlocks: Array<ItBlock>,
 };
 
+export const getASTfor = (file: string): BabylonFile => {
+  const data = readFileSync(file).toString();
+  const config = {plugins: ['*'], sourceType: 'module'};
+  return babylonParse(data, config);
+};
+
 export const parse = (file: string): BabylonParserResult => {
   const itBlocks: ItBlock[] = [];
   const expects: Expect[] = [];
 
-  const data = readFileSync(file).toString();
-
-  const config = {plugins: ['*'], sourceType: 'module'};
-  const ast = babylonParse(data, config);
+  const ast = getASTfor(file);
 
   // An `it`/`test` was found in the AST
   // So take the AST node and create an object for us
