@@ -22,13 +22,13 @@ import {
 } from 'jest-matcher-utils';
 import SnapshotState from './State';
 import {addSerializer, getSerializers} from './plugins';
-import {SNAPSHOT_EXTENSION} from './utils';
+import * as utils from './utils';
 
 const fileExists = (filePath: Path, hasteFS: HasteFS): boolean =>
   hasteFS.exists(filePath) || fs.existsSync(filePath);
 
 const cleanup = (hasteFS: HasteFS, update: SnapshotUpdateState) => {
-  const pattern = '\\.' + SNAPSHOT_EXTENSION + '$';
+  const pattern = '\\.' + utils.SNAPSHOT_EXTENSION + '$';
   const files = hasteFS.matchFiles(pattern);
   const filesRemoved = files
     .filter(
@@ -37,7 +37,7 @@ const cleanup = (hasteFS: HasteFS, update: SnapshotUpdateState) => {
           path.resolve(
             path.dirname(snapshotFile),
             '..',
-            path.basename(snapshotFile, '.' + SNAPSHOT_EXTENSION),
+            path.basename(snapshotFile, '.' + utils.SNAPSHOT_EXTENSION),
           ),
           hasteFS,
         ),
@@ -148,11 +148,12 @@ const toThrowErrorMatchingSnapshot = function(received: any, expected: void) {
 };
 
 module.exports = {
-  EXTENSION: SNAPSHOT_EXTENSION,
+  EXTENSION: utils.SNAPSHOT_EXTENSION,
   SnapshotState,
   addSerializer,
   cleanup,
   getSerializers,
   toMatchSnapshot,
   toThrowErrorMatchingSnapshot,
+  utils,
 };
